raynet_saas_optimization/backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
fastapi-cache2
redis
fastapi-limiter 